collabquest-backend/requirements.txt
# --- BACKEND CORE ---
fastapi
uvicorn[standard]
motor
beanie
pydantic
python-dotenv
requests
httpx

# --- AUTHENTICATION ---
python-jose
bcrypt
passlib
email-validator

# --- NEW AI STACK (OpenRouter / RAG) ---
openai
langchain
langgraph
langchain-community
langchain-openai
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers
pypdf
tiktoken

# --- LEGACY / GOOGLE STACK (Keep for Roadmap & OAuth) ---
google-auth-httplib2
google-generativeai
langchain-google-genai